+++ verb.el
@@ -426,7 +426,7 @@ If REMOVE is nil, add the necessary keywords to
         ["Export request to Verb" verb-export-request-on-point-verb]
         "--"
         ["Customize Verb" verb-customize-group]
-        ["View log" verb-view-log]))
+        ["Show log" verb-show-log]))
     map)
   "Keymap for `verb-mode'.")
 
@@ -590,12 +590,15 @@ KEY and VALUE must be strings.  KEY must not be the empty string."
     (define-key map (kbd "C-c C-r C-r") #'verb-toggle-show-headers)
     (define-key map (kbd "C-c C-r C-k") #'verb-kill-response-buffer-and-window)
     (define-key map (kbd "C-c C-r C-f") #'verb-re-send-request)
+    (define-key map (kbd "C-c C-r C-s") #'verb-show-request)
     (easy-menu-define verb-response-body-mode-menu map
       "Menu for Verb response body mode"
       '("Verb[Body]"
         ["Toggle show response headers" verb-toggle-show-headers]
         ["Kill buffer and window" verb-kill-response-buffer-and-window]
-        ["Re-send request" verb-re-send-request]))
+        "--"
+        ["Re-send request" verb-re-send-request]
+        ["Show corresponding request" verb-show-request]))
     map)
   "Keymap for `verb-response-body-mode'.")
 
@@ -614,10 +617,13 @@ KEY and VALUE must be strings.  KEY must not be the empty string."
             (verb-toggle-show-headers))))
     (setq header-line-format nil)))
 
-(defun verb-view-log ()
+(defun verb-show-log ()
   "Switch to the *Verb Log* buffer."
   (interactive)
   (switch-to-buffer (get-buffer-create verb--log-buffer-name)))
+
+;; Old function name (<2.11.0)
+(defalias 'verb-view-log 'verb-show-log)
 
 (defun verb-customize-group ()
   "Show the Customize menu buffer for the Verb package group."
@@ -1093,6 +1099,34 @@ Set the buffer's `verb-kill-this-buffer' variable locally to t."
         (insert key ": " value "\n")))
     (unless (zerop (buffer-size))
       (backward-delete-char 1))))
+
+(defun verb-show-request ()
+  "Show the corresponding HTTP request for a received response.
+The request will be displayed in a format very similar in which it was
+written originally, on a separate buffer.  Note that code tags will be
+shown with their values already evaluated.  Some of the headers sent
+by url.el may also differ from the headers shown with this function,
+see the project's README.md file for more information.  Return the
+buffer used to show the request."
+  (interactive)
+  (unless verb-response-body-mode
+    (user-error "%s" "This buffer is not showing an HTTP response"))
+
+  (let ((rs (oref verb-http-response request))
+        (name (buffer-name)))
+    (with-current-buffer (get-buffer-create "*Show HTTP Request*")
+      ;; Set up buffer
+      (unless (derived-mode-p 'org-mode)
+        (org-mode))
+      (verb--ensure-verb-mode)
+      (erase-buffer)
+
+      ;; Insert the request contents and show the buffer
+      (insert "* Corresponding HTTP request for response in "
+              name)
+      (newline)
+      (insert (verb-request-spec-to-string rs))
+      (switch-to-buffer-other-window (current-buffer)))))
 
 (defun verb-toggle-show-headers ()
   "Show or hide the HTTP response's headers on a separate buffer."